pull: detect root blob mediaType if missing (#189)
After downloading the root blob, which can be either a manifest or index, we need to decide which manifest kind we are working with.
The OCI spec actually doesn't require setting an explicit mediaType field, so we try to guess the correct type based on the JSON structure if no mediaType is set.

Closes #186