containerd image loading platform filtering
Currently when loading a image via containerd backend all blobs are loaded regardless of the selected platform making the load times take much longer.
Instead, we now only load the host platform by default. If a specific platform is given, we load it.
If a special "all" platform is given, we load a full index.