image_layer: avoid creating ".cas" directory
Previously, we had the idea to create a content-addressable storage across multiple layers. For that to work, we created a single location in the image (/.cas), where we would store files using their hash. It turns out that hardlinks only work reliably within layers (not across), so this content-addressable storage isn't needed. Instead, we now a file in the requested location (the first time we see it). If the file is written again, we hardlink any additional writes to the path of the first occurence.