image_layer: allow omitting runfiles (#261)
The new `include_runfiles` attribute allows users to disable runfiles in tar layers.
If set to True (default), any executable will be stored in the image along with its runfiles.
If set to False, the runfiles are omitted.
This can make things like static Go binaries smaller but may result in broken executables (when the runfiles are used at runtime).