pull: configurable Docker config path for lazy layer download
Lazy layer downloads run in Bazel's sandbox and by default, most environment variables are unset.
This leads to issues with authenticated container registries: We need to
locate the user's Docker config for authentication, but hides the
username, $HOME, etc.

With this change, users can set the location of the Docker config for
lazy image pulls in multiple ways:

* --@rules_img//img/settings:docker_config_path=/home/<USERNAME>/.docker/config.json
* --action_env=DOCKER_CONFIG=/home/<USERNAME>/.docker

Work towards #266