images.pull: Synchronize all missing facts in parallel
Since we are using mctx.execute to fetch manifests when building up a reference graph of blobs,
we cannot run multiple commands in parallel.
Before, this would make fetching slow when the lockfile was out of date.
We now use a single native Go command to sync the oci ref graph and
download all missing entries in parallel (in a single invocation of mctx.execute).

This will keep update times constant extremely short.