image_load: enable use of loader as data dependency in test
Currently, we automatically select a loader binary for the host, regardless of the "--platforms" setting.
That works great when it is used as a top-level target in "bazel run".
However, if "image_load" is used as part of an integration test,
we want the loader binary to be built for the target platform instead.
This change enables loading a container image at the beginning of an integration test.

Users need to decide what context image_load shall be used in:

- `tool_cfg = "host"` (default) works best with `bazel run`
- `tool_cfg = "target"` makes the loader usable as a `data` dependency of another target.